# Convert BigQuery field validation from a regex to hand-written code