--- a/apis/Google.Cloud.BigQuery.V2/Google.Cloud.BigQuery.V2/BigQueryNumeric.cs
+++ b/apis/Google.Cloud.BigQuery.V2/Google.Cloud.BigQuery.V2/BigQueryNumeric.cs
@@ -36,6 +36,7 @@ namespace Google.Cloud.BigQuery.V2
 
         private static readonly BigInteger s_integerScalingFactor = new BigInteger(1_000_000_000L);
         // TODO: Don't require a 0 before the decimal point.
+        // TODO: Replace with manual validation if we find this is a performance bottleneck (as it was with field name validation).
         private static readonly Regex s_validation = new Regex(@"^-?[0-9]+\.?[0-9]*$");
 
         // Note: the following properties must be declared *after* s_maxValue and s_minValue. Initialization order matters.