… apis/Google.Cloud.BigQuery.V2/Google.Cloud.BigQuery.V2/AssemblyInfo.cs