This is entirely for the sake of performance. Benchmark results on my machine:

netcoreapp2.0:
- Short: 180ns => 9ns
- Long: 939ns => 96ns

netcoreapp3.1:
- Short: 159ns => 8ns
- Long: 887ns => 79ns

net461:
- Short: 197ns => 8ns
- Long: 1425ns => 86ns

Fixes #4975